+++ x-pack/plugins/ml/public/application/trained_models/models_management/models_list.tsx
@@ -26,11 +26,10 @@ import { EuiTableSelectionType } from '@elastic/eui/src/components/basic_table/t
 import { Action } from '@elastic/eui/src/components/basic_table/action_types';
 import { FIELD_FORMAT_IDS } from '@kbn/field-formats-plugin/common';
 import { isPopulatedObject } from '@kbn/ml-is-populated-object';
-import { getUserInputThreadingParamsProvider } from './start_deployment_setup';
-import { getAnalysisType } from '../../data_frame_analytics/common';
+import { TrainedModelActions } from './model_actions';
 import { ModelsTableToConfigMapping } from '.';
 import { ModelsBarStats, StatsBar } from '../../components/stats_bar';
-import { useMlKibana, useMlLocator, useNavigateToPath, useTimefilter } from '../../contexts/kibana';
+import { useMlKibana, useTimefilter } from '../../contexts/kibana';
 import { useTrainedModelsApiService } from '../../services/ml_api_service/trained_models';
 import {
   ModelPipelines,
@@ -38,7 +37,6 @@ import {
   TrainedModelStat,
 } from '../../../../common/types/trained_models';
 import { BUILT_IN_MODEL_TAG } from '../../../../common/constants/data_frame_analytics';
-import { DataFrameAnalysisConfigType } from '../../../../common/types/data_frame_analytics';
 import { DeleteModelsModal } from './delete_models_modal';
 import { ML_PAGES } from '../../../../common/constants/locator';
 import { ListingPageUrlState } from '../../../../common/types/common';
@@ -48,14 +46,9 @@ import { useTableSettings } from '../../data_frame_analytics/pages/analytics_man
 import { useToastNotificationService } from '../../services/toast_notification_service';
 import { useFieldFormatter } from '../../contexts/kibana/use_field_formatter';
 import { useRefresh } from '../../routing/use_refresh';
-import {
-  BUILT_IN_MODEL_TYPE,
-  DEPLOYMENT_STATE,
-  TRAINED_MODEL_TYPE,
-} from '../../../../common/constants/trained_models';
-import { getUserConfirmationProvider } from './force_stop_dialog';
+import { BUILT_IN_MODEL_TYPE } from '../../../../common/constants/trained_models';
 import { SavedObjectsWarning } from '../../components/saved_objects_warning';
-import { isTestable, isTestEnabled, TestTrainedModelFlyout } from './test_models';
+import { TestTrainedModelFlyout } from './test_models';
 
 type Stats = Omit<TrainedModelStat, 'model_id'>;
 
@@ -85,15 +78,10 @@ export const ModelsList: FC<Props> = ({
 }) => {
   const {
     services: {
-      application: { navigateToUrl, capabilities },
-      overlays,
-      theme,
+      application: { capabilities },
       docLinks,
     },
   } = useMlKibana();
-  const urlLocator = useMlLocator()!;
-
-  const startModelDeploymentDocUrl = docLinks.links.ml.startTrainedModelsDeployment;
 
   useTimefilter({ timeRangeSelector: false, autoRefreshSelector: true });
 
@@ -117,12 +105,10 @@ export const ModelsList: FC<Props> = ({
   const searchQueryText = pageState.queryText ?? '';
 
   const canDeleteTrainedModels = capabilities.ml.canDeleteTrainedModels as boolean;
-  const canStartStopTrainedModels = capabilities.ml.canStartStopTrainedModels as boolean;
-  const canTestTrainedModels = capabilities.ml.canTestTrainedModels as boolean;
 
   const trainedModelsApiService = useTrainedModelsApiService();
 
-  const { displayErrorToast, displaySuccessToast } = useToastNotificationService();
+  const { displayErrorToast } = useToastNotificationService();
 
   const [isLoading, setIsLoading] = useState(false);
   const [items, setItems] = useState<ModelItem[]>([]);
@@ -132,14 +118,6 @@ export const ModelsList: FC<Props> = ({
     {}
   );
   const [showTestFlyout, setShowTestFlyout] = useState<ModelItem | null>(null);
-  const getUserConfirmation = useMemo(() => getUserConfirmationProvider(overlays, theme), []);
-
-  const getUserInputThreadingParams = useMemo(
-    () => getUserInputThreadingParamsProvider(overlays, theme.theme$, startModelDeploymentDocUrl),
-    [overlays, theme.theme$, startModelDeploymentDocUrl]
-  );
-
-  const navigateToPath = useNavigateToPath();
 
   const isBuiltInModel = useCallback(
     (item: ModelItem) => item.tags.includes(BUILT_IN_MODEL_TAG),
@@ -279,211 +257,20 @@ export const ModelsList: FC<Props> = ({
       }));
   }, [items]);
 
-  /**
-   * Table actions
-   */
   const actions: Array<Action<ModelItem>> = [
     {
-      name: i18n.translate('xpack.ml.trainedModels.modelsList.viewTrainingDataActionLabel', {
-        defaultMessage: 'View training data',
-      }),
-      description: i18n.translate('xpack.ml.trainedModels.modelsList.viewTrainingDataActionLabel', {
-        defaultMessage: 'View training data',
-      }),
-      icon: 'visTable',
-      type: 'icon',
-      available: (item) => !!item.metadata?.analytics_config?.id,
-      onClick: async (item) => {
-        if (item.metadata?.analytics_config === undefined) return;
-
-        const analysisType = getAnalysisType(
-          item.metadata?.analytics_config.analysis
-        ) as DataFrameAnalysisConfigType;
-
-        const url = await urlLocator.getUrl({
-          page: ML_PAGES.DATA_FRAME_ANALYTICS_EXPLORATION,
-          pageState: {
-            jobId: item.metadata?.analytics_config.id as string,
-            analysisType,
-            ...(analysisType === 'classification' || analysisType === 'regression'
-              ? {
-                  queryText: `${item.metadata?.analytics_config.dest.results_field}.is_training : true`,
-                }
-              : {}),
-          },
-        });
-
-        await navigateToUrl(url);
-      },
-      isPrimary: true,
-    },
-    {
-      name: i18n.translate('xpack.ml.inference.modelsList.analyticsMapActionLabel', {
-        defaultMessage: 'Analytics map',
-      }),
-      description: i18n.translate('xpack.ml.inference.modelsList.analyticsMapActionLabel', {
-        defaultMessage: 'Analytics map',
-      }),
-      icon: 'graphApp',
-      type: 'icon',
-      isPrimary: true,
-      available: (item) => !!item.metadata?.analytics_config?.id,
-      onClick: async (item) => {
-        const path = await urlLocator.getUrl({
-          page: ML_PAGES.DATA_FRAME_ANALYTICS_MAP,
-          pageState: { modelId: item.model_id },
-        });
-
-        await navigateToPath(path, false);
-      },
-    },
-    {
-      name: i18n.translate('xpack.ml.inference.modelsList.startModelDeploymentActionLabel', {
-        defaultMessage: 'Start deployment',
-      }),
-      description: i18n.translate('xpack.ml.inference.modelsList.startModelDeploymentActionLabel', {
-        defaultMessage: 'Start deployment',
-      }),
-      icon: 'play',
-      type: 'icon',
-      isPrimary: true,
-      enabled: (item) => {
-        const { state } = item.stats?.deployment_stats ?? {};
+      render: (item) => {
         return (
-          canStartStopTrainedModels &&
-          !isLoading &&
-          state !== DEPLOYMENT_STATE.STARTED &&
-          state !== DEPLOYMENT_STATE.STARTING
+          <TrainedModelActions
+            modelItem={item}
+            setShowTestFlyout={setShowTestFlyout}
+            isLoading={isLoading}
+            setIsLoading={setIsLoading}
+            setModelIdsToDelete={setModelIdsToDelete}
+            onRefresh={fetchModelsData}
+          />
         );
       },
-      available: (item) => item.model_type === TRAINED_MODEL_TYPE.PYTORCH,
-      onClick: async (item) => {
-        const threadingParams = await getUserInputThreadingParams(item.model_id);
-
-        if (!threadingParams) return;
-
-        try {
-          setIsLoading(true);
-          await trainedModelsApiService.startModelAllocation(item.model_id, {
-            number_of_allocations: threadingParams.numOfAllocations,
-            threads_per_allocation: threadingParams.threadsPerAllocations,
-          });
-          displaySuccessToast(
-            i18n.translate('xpack.ml.trainedModels.modelsList.startSuccess', {
-              defaultMessage: 'Deployment for "{modelId}" has been started successfully.',
-              values: {
-                modelId: item.model_id,
-              },
-            })
-          );
-          await fetchModelsData();
-        } catch (e) {
-          displayErrorToast(
-            e,
-            i18n.translate('xpack.ml.trainedModels.modelsList.startFailed', {
-              defaultMessage: 'Failed to start "{modelId}"',
-              values: {
-                modelId: item.model_id,
-              },
-            })
-          );
-          setIsLoading(false);
-        }
-      },
-    },
-    {
-      name: i18n.translate('xpack.ml.inference.modelsList.stopModelDeploymentActionLabel', {
-        defaultMessage: 'Stop deployment',
-      }),
-      description: i18n.translate('xpack.ml.inference.modelsList.stopModelDeploymentActionLabel', {
-        defaultMessage: 'Stop deployment',
-      }),
-      icon: 'stop',
-      type: 'icon',
-      isPrimary: true,
-      available: (item) => item.model_type === TRAINED_MODEL_TYPE.PYTORCH,
-      enabled: (item) =>
-        canStartStopTrainedModels &&
-        !isLoading &&
-        isPopulatedObject(item.stats?.deployment_stats) &&
-        item.stats?.deployment_stats?.state !== DEPLOYMENT_STATE.STOPPING,
-      onClick: async (item) => {
-        const requireForceStop = isPopulatedObject(item.pipelines);
-
-        if (requireForceStop) {
-          const hasUserApproved = await getUserConfirmation(item);
-          if (!hasUserApproved) return;
-        }
-
-        if (requireForceStop) {
-          const hasUserApproved = await getUserConfirmation(item);
-          if (!hasUserApproved) return;
-        }
-
-        try {
-          setIsLoading(true);
-          await trainedModelsApiService.stopModelAllocation(item.model_id, {
-            force: requireForceStop,
-          });
-          displaySuccessToast(
-            i18n.translate('xpack.ml.trainedModels.modelsList.stopSuccess', {
-              defaultMessage: 'Deployment for "{modelId}" has been stopped successfully.',
-              values: {
-                modelId: item.model_id,
-              },
-            })
-          );
-          // Need to fetch model state updates
-          await fetchModelsData();
-        } catch (e) {
-          displayErrorToast(
-            e,
-            i18n.translate('xpack.ml.trainedModels.modelsList.stopFailed', {
-              defaultMessage: 'Failed to stop "{modelId}"',
-              values: {
-                modelId: item.model_id,
-              },
-            })
-          );
-          setIsLoading(false);
-        }
-      },
-    },
-    {
-      name: i18n.translate('xpack.ml.trainedModels.modelsList.deleteModelActionLabel', {
-        defaultMessage: 'Delete model',
-      }),
-      description: i18n.translate('xpack.ml.trainedModels.modelsList.deleteModelActionLabel', {
-        defaultMessage: 'Delete model',
-      }),
-      'data-test-subj': 'mlModelsTableRowDeleteAction',
-      icon: 'trash',
-      type: 'icon',
-      color: 'danger',
-      isPrimary: false,
-      onClick: (model) => {
-        setModelIdsToDelete([model.model_id]);
-      },
-      available: (item) => canDeleteTrainedModels && !isBuiltInModel(item),
-      enabled: (item) => {
-        // TODO check for permissions to delete ingest pipelines.
-        // ATM undefined means pipelines fetch failed server-side.
-        return !isPopulatedObject(item.pipelines);
-      },
-    },
-    {
-      name: i18n.translate('xpack.ml.inference.modelsList.testModelActionLabel', {
-        defaultMessage: 'Test model',
-      }),
-      description: i18n.translate('xpack.ml.inference.modelsList.testModelActionLabel', {
-        defaultMessage: 'Test model',
-      }),
-      icon: 'inputOutput',
-      type: 'icon',
-      isPrimary: true,
-      available: isTestable,
-      onClick: setShowTestFlyout,
-      enabled: (item) => canTestTrainedModels && isTestEnabled(item),
     },
   ];
